factor our builder and related classes to dto package
