+++ server/src/main/java/org/tctalent/server/api/admin/CandidateAdminApi.java
@@ -44,6 +44,9 @@ import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 import org.springframework.web.reactive.function.client.WebClientException;
 import org.tctalent.anonymization.model.RegisterCandidate201Response;
+import org.tctalent.server.api.dto.CandidateBuilderSelector;
+import org.tctalent.server.api.dto.CandidateIntakeDataBuilderSelector;
+import org.tctalent.server.api.dto.DtoType;
 import org.tctalent.server.exception.ExportFailedException;
 import org.tctalent.server.exception.NoSuchObjectException;
 import org.tctalent.server.exception.SalesforceException;
