+++ server/src/main/java/org/tctalent/server/api/dto/CandidateBuilderSelector.java
@@ -14,7 +14,7 @@
  * along with this program. If not, see https://www.gnu.org/licenses/.
  */
 
-package org.tctalent.server.api.admin;
+package org.tctalent.server.api.dto;
 
 import java.util.Arrays;
 import java.util.HashSet;
